Added hosted sync/async document example files
Why is this change needed?
--------------------------
The way hosted documents work is different than how DocRaptor has worked
in the past. Since we've added a new method to help create hosted
sync/async documents we should add an example to portray how that works.

How does it address the issue?
------------------------------
This creates two new examples that create hosted documents and saves them
to a temporary file.

Note:
- We don't allow the example api key to create hosted documents so
  user's will need to provide their own if they want this to work.