Update strict param in docraptor.yaml
Why is this change needed?
--------------------------
The strict parameter only allowed a single value of "none" when valid
values for it are "none" and "html"

How does it address the issue?
------------------------------
Strict now allows "none" and "html" as values.
Since it's not a required parameter we removed the default value.

We also fixed the valid values for the strict parameter to match the api.
https://docraptor.com/documentation/api#api_strict

Thirdly we moved pipeline down after the required fields so the important
fields are listed first.

Co-authored-by: Alex Overbeck <alex@expectedbehavior.com>