Switch to openapi/docker for client generation
Why is this change needed?
--------------------------
We want to generate new agents in order to switch API endpoints. Swagger
is old and busted.

Environment slog is a major hassle. Sometimes bit rot sets in between
agent improvements.

How does it address the issue?
------------------------------
Switch to openapi-generator because it's better maintained than swagger.
Also use their docker offering so we don't need to deal with the nuances
of their java setup. This essentially gives us a fully packaged binary
that we can use to generate client code.

We did this with the PHP agent previously and it was easier than the old
java setup with swagger.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?