Update CHANGELOG to specify Java version support
Why is this change needed?
--------------------------
We want to clarify what versions we support and test.

How does it address the issue?
------------------------------
Add a CHANGELOG entry. Also add a Java 9 and 10 testing to complete the
set.

This also updates the `.runtime-environments` file to order images from
least specific and newest to oldest. The newer less specific versions
should automatically update to new versions of Java whenever we run the
tests which should give us indicators if things still work. When new
versions are released we should add those images to the list so each
version gets tested.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?