Handle `docker pull` network issues
Why is this change needed?
--------------------------
If `docker pull` can't contact the registry for some reason you can get
an error. This doesn't need to break things if you already have some
version of the container image.

How does it address the issue?
------------------------------
Perform the pull separately, and don't error if it fails. This will
update container images like `mono:latest` so we can easily keep track
of whether or not the latest images fail, but it won't stop the test
process if there are temporary connection issues to the registry.

Usually you'll have a previous version of a container image stored
locally, and that'll be used by `docker run`.

We print a message in case you *really* want to run against the latest
image.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?