Add patch to fix publishing issue in the future
Why is this change needed?
--------------------------
See the patch or the previous commit for the details.

How does it address the issue?
------------------------------
This encapsulates the change that is needed after each client generation
to allow publishing.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?