Improve release process
Why is this change needed?
--------------------------
Ideally releasing uses docker to prevent environment issues.

How does it address the issue?
------------------------------
Add `script/release` to allowing scripting the process.

This includes facilitating the setup of a gpg key for maven publishing.

This whole process has been difficult to remember and execute correctly
every time, so this scripts as much as is reasonable to help the
developer along.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778
https://stackoverflow.com/questions/57591432/gpg-signing-failed-inappropriate-ioctl-for-device-on-macos-with-maven
https://www.gnupg.org/documentation/manuals/gnupg/GPG-Configuration.html
https://github.com/keybase/keybase-issues/issues/2798
https://askubuntu.com/questions/29889/how-do-i-check-if-my-openpgp-key-is-in-the-ubuntu-keyserver
https://stackoverflow.com/questions/51504367/gpg-agent-forwarding-inappropriate-ioctl-for-device
https://superuser.com/questions/1150165/get-list-of-secret-key-ids
https://serverfault.com/questions/691120/how-to-generate-gpg-key-without-user-interaction

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?