Improve generated file review
Why is this change needed?
--------------------------
These files might have useful changes that we want to review.

How does it address the issue?
------------------------------
Don't remove these since we'll want to review the diffs between the old
and new versions to see if we should make any updates to non-generated
code.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29007795/todolists/5257298600

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?