Add a new test for prince_options[insecure]=true
Why is this change needed?
--------------------------
Csharp had an issue with PrinceOptions, so I made a test. Since the
basic test exists I'll copy it to the other clients to verify everything
continues to work in the future.

How does it address the issue?
------------------------------
Add a test for PrinceOptions.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29007795/todolists/5257298600

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?